[REFACTOR] Separate locations and font logic from device
