+++ packages/devices/src/index.ts
@@ -2,4 +2,6 @@ export * from "./DeviceContext";
 export * from "./DeviceWrapper";
 export * from "./useDevice";
 export * from "./getDevices";
+export * from "./getFonts";
+export * from "./getLocations";
 export type { DeviceSelections } from "./types";
