+++ packages/devices/tsconfig.json
@@ -1,6 +1,6 @@
 {
     "extends": "../../tsconfig.json",
-    "include": ["src/**/*"],
+    "include": ["src/**/*", "src/locations.json"],
     "compilerOptions": {
         "outDir": "./dist",
         "rootDir": "./src",
